backend/apps/testimonials/views.py
from rest_framework import viewsets
from rest_framework.permissions import IsAdminUser, AllowAny
from .models import Testimonial
from .serializers import TestimonialSerializer


class TestimonialViewSet(viewsets.ModelViewSet):
    queryset = Testimonial.objects.all()
    serializer_class = TestimonialSerializer

    def get_permissions(self):
        # Allow read-only access to anyone; write actions require admin
        if self.request.method in ("GET", "HEAD", "OPTIONS"):
            return [AllowAny()]
        return [IsAdminUser()]
