backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
drf-spectacular==0.27.2
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
stripe>=7.0.0
requests>=2.28.0
reportlab>=4.0.0
